+++ components/places/src/api/places_api.rs
@@ -54,11 +54,17 @@ pub fn get_registered_sync_engine(engine_id: &SyncEngineId) -> Option<Box<dyn Sy
         Some(places_api) => match create_sync_engine(&places_api, engine_id) {
             Ok(engine) => Some(engine),
             Err(e) => {
-                error_support::report_error!(
-                    "places-no-registered-sync-engine",
-                    "places: get_registered_sync_engine: {}",
-                    e
-                );
+                // Report this to Sentry, except if it's an open database error.  That indicates
+                // that there is a registered sync engine, but the connection is busy so we can't
+                // open it.  This is a known issue that we don't need more reports for (see
+                // https://github.com/mozilla/application-services/issues/5237 for discussion).
+                if !matches!(e, Error::OpenDatabaseError(_)) {
+                    error_support::report_error!(
+                        "places-no-registered-sync-engine",
+                        "places: get_registered_sync_engine: {}",
+                        e
+                    );
+                }
                 None
             }
         },
@@ -124,7 +130,7 @@ pub struct SyncState {
 /// For uniffi we need to expose our `Arc` returning constructor as a global function :(
 /// https://github.com/mozilla/uniffi-rs/pull/1063 would fix this, but got some pushback
 /// meaning we are forced into this unfortunate workaround.
-#[handle_error]
+#[handle_error(crate::Error)]
 pub fn places_api_new(db_name: impl AsRef<Path>) -> ApiResult<Arc<PlacesApi>> {
     PlacesApi::new(db_name)
 }
@@ -451,7 +457,7 @@ impl PlacesApi {
         Ok(())
     }
 
-    #[handle_error]
+    #[handle_error(crate::Error)]
     pub fn reset_history(&self) -> ApiResult<()> {
         // Take the lock to prevent syncing while we're doing this.
         let _guard = self.sync_state.lock();