+++ megazords/ios-rust/src/lib.rs
@@ -13,7 +13,11 @@ pub use logins;
 pub use nimbus;
 pub use places;
 pub use push;
+// TODO: Drop this dependency once firefox-ios switches to using `rust_log_forwarder` for log
+// forwarding.
 pub use rc_log_ffi;
+pub use rust_log_forwarder;
 pub use sync15;
+pub use sync_manager;
 pub use tabs;
 pub use viaduct_reqwest;