+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+# v97.1.0 (_2023-02-24_)
+
+[Full Changelog](https://github.com/mozilla/application-services/compare/v97.0.0...v97.1.0)
+
+## Tabs
+
+### 🦊 What's Changed 🦊
+
+- The Tabs engine now trims the payload to be under the max the server will accept ([#5376](https://github.com/mozilla/application-services/pull/5376))
+
+
+## Sync Manager
+
+### 🦊 What's Changed 🦊
+
+- Exposing the Sync Manager component to iOS by addressing the existing naming collisions, adding logic to process the telemetry
+  data returned in the component's `sync` function, and adding the component to the iOS megazord ([#5359](https://github.com/mozilla/application-services/pull/5359)).
+
+# v97.0.0 (_2023-02-22_)
+
+[Full Changelog](https://github.com/mozilla/application-services/compare/v96.4.0...v97.0.0)
+
+## Nimbus ⛅️🔬🔭
+### 🦊 What's Changed 🦊
+- Updated the Nimbus Gradle Plugin to fix a number of issues after migrating it to this repository ([#5348](https://github.com/mozilla/application-services/pull/5348))
+- Good fences: protected calls out to the error reporter with a `try`/`catch` ([#5366](https://github.com/mozilla/application-services/pull/5366))
+- Updated the Nimbus FML CLI to only import the R class if it will be used by a feature property ([#5361](https://github.com/mozilla/application-services/pull/5361))
+### ⚠️ Breaking Changes ⚠️
+  - Android and iOS: Several errors have been moved to an internal support library and will no longer be reported as top-level Nimbus errors. They should still be accessible through `NimbusError.ClientError`. They are: `RequestError`, `ResponseError`, and `BackoffError`. ([#5369](https://github.com/mozilla/application-services/pull/5369))
+
 # v96.4.0 (_2023-01-30_)
 
 [Full Changelog](https://github.com/mozilla/application-services/compare/v96.3.0...v96.4.0)